backend/requirements.txt
ansible==2.8.4
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
entrypoints==0.3
flake8==3.7.8
future==0.17.1
gunicorn==19.9.0
importlib-metadata==0.20
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
